Move Notice to Common/UI
