+++ UI/OnScreenDisplay.h
@@ -51,39 +51,3 @@ private:
 	OnScreenMessagesView *osmView_ = nullptr;
 };
 
-enum class NoticeLevel {
-	SUCCESS,
-	INFO,
-	WARN,
-	ERROR,
-};
-
-class NoticeView : public UI::InertView {
-public:
-	NoticeView(NoticeLevel level, std::string_view text, std::string_view detailsText, UI::LayoutParams *layoutParams = 0)
-		: InertView(layoutParams), level_(level), text_(text), detailsText_(detailsText), iconName_("") {}
-
-	void SetIconName(std::string_view name) {
-		iconName_ = name;
-	}
-	void SetText(std::string_view text) {
-		text_ = text;
-	}
-	void SetLevel(NoticeLevel level) {
-		level_ = level;
-	}
-	void SetSquishy(bool squishy) {
-		squishy_ = squishy;
-	}
-
-	void GetContentDimensionsBySpec(const UIContext &dc, UI::MeasureSpec horiz, UI::MeasureSpec vert, float &w, float &h) const override;
-	void Draw(UIContext &dc) override;
-
-private:
-	std::string text_;
-	std::string detailsText_;
-	std::string iconName_;
-	NoticeLevel level_;
-	mutable float height1_ = 0.0f;
-	bool squishy_ = false;
-};
