+++ UI/PauseScreen.cpp
@@ -24,6 +24,7 @@
 #include "Common/UI/Context.h"
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/PopupScreens.h"
+#include "Common/UI/Notice.h"
 #include "Common/GPU/thin3d.h"
 
 #include "Common/Data/Text/I18n.h"
