+++ UI/RemoteISOScreen.cpp
@@ -36,6 +36,7 @@
 
 #include "Common/File/PathBrowser.h"
 #include "Common/UI/PopupScreens.h"
+#include "Common/UI/Notice.h"
 #include "Common/Data/Format/JSONReader.h"
 #include "Common/Data/Text/I18n.h"
 #include "Common/Common.h"
