+++ UI/SystemInfoScreen.cpp
@@ -18,6 +18,7 @@
 #include "Common/Render/Text/draw_text.h"
 #include "Common/System/Request.h"
 #include "Common/UI/Context.h"
+#include "Common/UI/Notice.h"
 #include "Core/System.h"
 #include "Core/Config.h"
 #include "GPU/GPUState.h"  // ugh
